+++ multicore-sys-monitor@ccadeptic23/files/multicore-sys-monitor@ccadeptic23/5.6/applet.js
@@ -15,25 +15,34 @@ const windowTracker = imports.gi.Cinnamon.WindowTracker.get_default();
 const Util = imports.misc.util;
 const St = imports.gi.St;
 const Gtk = imports.gi.Gtk;
-const {
-  reloadExtension,
-  Type
-} = imports.ui.extension; //Extension
+const Extension = imports.ui.extension;
+//~ const {
+  //~ reloadExtension,
+  //~ Type
+//~ } = imports.ui.extension; //Extension
 
-const { to_string } = require("./lib/tostring");
-const { readFileAsync } = require("./lib/readFileAsync");
-const Graphs = require("./lib/Graphs");
-const {
+function _require(relPath) {
+    if (Extension.getCurrentExtension) {
+        var Me = Extension.getCurrentExtension();
+        return Me.imports[relPath];
+    } else {
+        return require(relPath);
+    }
+}
+var { to_string } = _require("./lib/tostring");
+var { readFileAsync } = _require("./lib/readFileAsync");
+var Graphs = _require("./lib/Graphs");
+var {
   timeout_add,
   setTimeout,
   clearTimeout,
   source_exists,
   source_remove,
-  remove_all_sources } = require("./lib/mainloopTools");
+  remove_all_sources } = _require("./lib/mainloopTools");
 
 const UUID = "multicore-sys-monitor@ccadeptic23";
 const HOME_DIR = GLib.get_home_dir();
-const APPLET_DIR = HOME_DIR + "/.local/share/cinnamon/applets/" + UUID;
+const APPLET_DIR = GLib.get_user_data_dir() + "/cinnamon/applets/" + UUID;
 const PATH2SCRIPTS = APPLET_DIR + "/scripts";
 const XDG_RUNTIME_DIR = GLib.getenv("XDG_RUNTIME_DIR");
 const NETWORK_DEVICES_STATUS_PATH = XDG_RUNTIME_DIR + "/network_devices";
@@ -220,6 +229,7 @@ class MCSM extends Applet.TextIconApplet {
         this.settings.bind("graphStep", "graphStep");
         this.settings.bind("graphSpacing", "graphSpacing");
         this.settings.bind("CPU_groupInSetsOf4", "CPU_groupInSetsOf4");
+        this.settings.bind("quickTooltip", "quickTooltip");
         this.settings.bind("percentAtEndOfLine", "percentAtEndOfLine");
         this.settings.bind("tooltipFontSize", "tooltipFontSize", () => { this.setTooltipStyle(); });
         this.settings.bind("tooltipChangeBackgroundColor", "tooltipChangeBackgroundColor", () => { this.setTooltipStyle(); });
@@ -267,6 +277,7 @@ class MCSM extends Applet.TextIconApplet {
         this.settings.bind("CPU_squared", "CPU_squared");
 
         this.settings.bind("CPU_showTemp", "CPU_showTemp");
+        this.settings.bind("CPU_showTempUnit", "CPU_showTempUnit");
         this.settings.bind("CPU_temp_hovering_only", "CPU_temp_hovering_only");
         this.settings.bind("CPU_tempInFahrenheit", "CPU_tempInFahrenheit");
         this.settings.bind("CPU_tempPath", "CPU_tempPath");
@@ -484,16 +495,19 @@ class MCSM extends Applet.TextIconApplet {
         this._applet_label.connect("motion-event", (actor, event) => {
             this.hovered = true;
             this._applet_tooltip.preventShow = false;
+            if (!this.quickTooltip) return;
             this._applet_tooltip.show()
         });
         this.actor.connect("motion-event", (actor, event) => {
             this.hovered = true;
             this._applet_tooltip.preventShow = false;
+            if (!this.quickTooltip) return;
             this._applet_tooltip.show()
         });
         this._applet_tooltip._onMotionEvent = function(actor, event) {
             this.hovered = true;
             this._applet_tooltip.preventShow = false;
+            if (!this.quickTooltip) return;
             this._applet_tooltip.show()
         };
         this.actor.connect("enter-event", (actor, event) => {
@@ -506,6 +520,7 @@ class MCSM extends Applet.TextIconApplet {
             }
             this._applet_tooltip.preventShow = false;
             this._setTooltip();
+            if (!this.quickTooltip) return;
             this._applet_tooltip.show();
         });
         this.actor.connect("leave-event", (actor, event) => {
@@ -931,18 +946,24 @@ class MCSM extends Applet.TextIconApplet {
           var intervalId = null;
           intervalId = setTimeout(() => {
                 clearTimeout(intervalId);
+                this._removeEnlightenment();
                 app = windowTracker.get_app_from_pid(pid);
                 if (app != null) {
                     let window = app.get_windows()[0];
                     this.settingsTab = tab;
 
-                    window.maximize(VERTICAL);
+                    //~ window.maximize(VERTICAL);
                     window.activate(300);
+                    if (window.set_maximize_flags) {
+                        window.set_maximize_flags(VERTICAL);
+                        window.maximize();
+                    } else {
+                        window.maximize(VERTICAL);
+                    }
                     this.settingsWindow = window;
                     app.connect("windows-changed", () => { this.settingsWindow = undefined; });
-                    this._removeEnlightenment();
                 }
-            }, 600);
+            }, 1000);
         }
         // Returns the pid:
         return pid;
@@ -1757,7 +1778,7 @@ class MCSM extends Applet.TextIconApplet {
     }
 
     refreshAll() {
-        reloadExtension(UUID, Type.APPLET);
+        Extension.reloadExtension(UUID, Extension.Type.APPLET);
     }
 
     on_panel_height_changed() {
@@ -2097,7 +2118,9 @@ class MultiCpuDataProvider {
             trans += " (" + this.applet.CPU_loadAverage.toString().padStart(2) + " %)";
         if (this.applet.CPU_showTemp && this.applet.CPU_temperature) {
             let temperature = 1 * this.applet.CPU_temperature;
-            let degree = (this.applet.CPU_tempInFahrenheit) ?  "°F" : "°C";
+            let degree = "°";
+            if (this.applet.CPU_showTempUnit)
+                degree = (this.applet.CPU_tempInFahrenheit) ?  "°F" : "°C";
             temperature = "" + temperature + degree;
             trans += " " + temperature;
         }